power: avoid implicit casts of int64_t to double
In powerd/system, as suggested by clang-tidy, avoid implicit of int64_t to
double. Mostly, the best, simple change is to make the casts explicit.

While arguably stylistically slightly better, this change is effectively
a no-op.

BUG=b:258053283
TEST=Local Tests + CQ

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4023518
Commit-Queue: Alan Green <avg@chromium.org>
Tested-by: Alan Green <avg@chromium.org>

+++ power_manager/powerd/system/internal_backlight.cc
@@ -200,9 +200,9 @@ void InternalBacklight::HandleTransitionTimeout() {
     double transition_fraction =
         (now - transition_start_time_).InMillisecondsF() /
         (transition_end_time_ - transition_start_time_).InMillisecondsF();
-    int64_t intermediate_amount =
-        lround(transition_fraction *
-               (transition_end_level_ - transition_start_level_));
+    int64_t intermediate_amount = lround(
+        transition_fraction *
+        static_cast<double>((transition_end_level_ - transition_start_level_)));
     new_level = transition_start_level_ + intermediate_amount;
   }
